Implemented basic database implementation on Order and OrderProduct classes, and moved the dtos to the schema folder.
